Billing amount should be float

+++ pureport/client/model_billing_plan.go
@@ -12,9 +12,9 @@ package client
 
 // BillingPlan struct for BillingPlan
 type BillingPlan struct {
-	Amount          int64  `json:"amount,omitempty"`
-	BillingInterval string `json:"billingInterval,omitempty"`
-	Id              string `json:"id,omitempty"`
-	SetupAmount     int64  `json:"setupAmount,omitempty"`
-	Term            string `json:"term,omitempty"`
+	Amount          float64 `json:"amount,omitempty"`
+	BillingInterval string  `json:"billingInterval,omitempty"`
+	Id              string  `json:"id,omitempty"`
+	SetupAmount     int64   `json:"setupAmount,omitempty"`
+	Term            string  `json:"term,omitempty"`
 }
